+++ src/DataType.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
+use Exception;
 use Stringable;
 
 enum DataType: string
@@ -15,25 +16,9 @@ enum DataType: string
     case Item = 'item';
 
     /**
-     * @throws SyntaxError
+     * @throws SyntaxError|Exception
      */
-    public function fromRfc9651(Stringable|string $httpValue): OuterList|InnerList|Parameters|Dictionary|Item
-    {
-        return $this->parse($httpValue, Ietf::Rfc9651);
-    }
-
-    /**
-     * @throws SyntaxError
-     */
-    public function fromRfc8941(Stringable|string $httpValue): OuterList|InnerList|Parameters|Dictionary|Item
-    {
-        return $this->parse($httpValue, Ietf::Rfc8941);
-    }
-
-    /**
-     * @throws SyntaxError
-     */
-    public function parse(Stringable|string $httpValue, ?Ietf $rfc = null): OuterList|InnerList|Parameters|Dictionary|Item
+    public function parse(Stringable|string $httpValue, ?Ietf $rfc = Ietf::Rfc9651): OuterList|InnerList|Parameters|Dictionary|Item
     {
         return match ($this) {
             self::List => OuterList::fromHttpValue($httpValue, $rfc),
@@ -45,7 +30,29 @@ enum DataType: string
     }
 
     /**
-     * @throws SyntaxError
+     * @throws SyntaxError|Exception
+     */
+    public function serialize(iterable $data, ?Ietf $rfc = Ietf::Rfc9651): string
+    {
+        return (match ($this) {
+            self::List => OuterList::fromPairs($data),
+            self::Dictionary => Dictionary::fromPairs($data),
+            self::Item => Item::fromPair([...$data]),
+            self::InnerList => InnerList::fromPair([...$data]),
+            self::Parameters => Parameters::fromPairs($data),
+        })->toHttpValue($rfc);
+    }
+
+    /**
+     * @throws SyntaxError|Exception
+     */
+    public function fromRfc9651(Stringable|string $httpValue): OuterList|InnerList|Parameters|Dictionary|Item
+    {
+        return $this->parse($httpValue, Ietf::Rfc9651);
+    }
+
+    /**
+     * @throws SyntaxError|Exception
      */
     public function toRfc9651(iterable $data): string
     {
@@ -53,32 +60,18 @@ enum DataType: string
     }
 
     /**
-     * @throws SyntaxError
+     * @throws SyntaxError|Exception
+     */
+    public function fromRfc8941(Stringable|string $httpValue): OuterList|InnerList|Parameters|Dictionary|Item
+    {
+        return $this->parse($httpValue, Ietf::Rfc8941);
+    }
+
+    /**
+     * @throws SyntaxError|Exception
      */
     public function toRfc8941(iterable $data): string
     {
         return $this->serialize($data, Ietf::Rfc8941);
-    }
-
-    /**
-     * @throws SyntaxError
-     */
-    public function serialize(iterable $data, ?Ietf $rfc = null): string
-    {
-        return $this->create($data)->toHttpValue($rfc);
-    }
-
-    /**
-     * @throws SyntaxError
-     */
-    public function create(iterable $data): OuterList|InnerList|Parameters|Dictionary|Item
-    {
-        return match ($this) {
-            self::List => OuterList::fromPairs($data),
-            self::Dictionary => Dictionary::fromPairs($data),
-            self::Item => Item::fromPair([...$data]),
-            self::InnerList => InnerList::fromPair([...$data]),
-            self::Parameters => Parameters::fromPairs($data),
-        };
     }
 }