+++ src/Type.php
@@ -11,6 +11,9 @@ use DateTimeInterface;
  */
 enum Type: string
 {
+    public const MAXIMUM_INT = 999_999_999_999_999;
+    public const MAXIMUM_FLOAT = 999_999_999_999;
+
     case Integer = 'integer';
     case Decimal = 'decimal';
     case String = 'string';
@@ -40,29 +43,29 @@ enum Type: string
     }
 
     /**
-     * @throws InvalidArgument if the value can not be resolved into a supported HTTP structured field data type
+     * @throws SyntaxError if the value can not be resolved into a supported HTTP structured field data type
      */
     public static function fromVariable(mixed $value): self
     {
-        return self::tryFromVariable($value) ?? throw new InvalidArgument(match (true) {
-            $value instanceof DateTimeInterface => 'The date is outside of the valid range for a HTTP structured field date type',
-            is_int($value) => 'The integer is outside of the valid range for a HTTP structured field integer type',
-            is_float($value) => 'The float is outside of the valid range for a HTTP structured field decimal type',
+        return self::tryFromVariable($value) ?? throw new SyntaxError(match (true) {
+            $value instanceof DateTimeInterface => 'The integer representation of a date is limited to 15 digits for a HTTP structured field date type.',
+            is_int($value) => 'The integer is limited to 15 digits for a HTTP structured field integer type.',
+            is_float($value) => 'The integer portion of decimals is limited to 12 digits for a HTTP structured field decimal type.',
             is_string($value) => 'The string contains characters that are invalid for a HTTP structured field string type',
             default => (is_object($value) ? 'An instance of "'.$value::class.'"' : 'A value of type "'.gettype($value).'"').' can not be used as an HTTP structured field value type.',
         });
     }
 
-    public static function tryFromVariable(mixed $variable): self|null
+    public static function tryFromVariable(mixed $variable): ?self
     {
         return match (true) {
             $variable instanceof Item,
             $variable instanceof Token,
             $variable instanceof DisplayString,
             $variable instanceof Bytes => $variable->type(),
-            $variable instanceof DateTimeInterface && 999_999_999_999_999 >= abs($variable->getTimestamp()) => Type::Date,
-            is_int($variable) && 999_999_999_999_999 >= abs($variable) => Type::Integer,
-            is_float($variable) && 999_999_999_999 >= abs(floor($variable)) => Type::Decimal,
+            $variable instanceof DateTimeInterface && self::MAXIMUM_INT >= abs($variable->getTimestamp()) => Type::Date,
+            is_int($variable) && self::MAXIMUM_INT >= abs($variable) => Type::Integer,
+            is_float($variable) && self::MAXIMUM_FLOAT >= abs(floor($variable)) => Type::Decimal,
             is_bool($variable) => Type::Boolean,
             is_string($variable) && 1 !== preg_match('/[^\x20-\x7f]/', $variable) => Type::String,
             default => null,
@@ -71,8 +74,6 @@ enum Type: string
 
     public function supports(mixed $value): bool
     {
-        $new = self::tryFromVariable($value);
-
-        return null !== $new && $new->equals($this);
+        return self::tryFromVariable($value)?->equals($this) ?? false;
     }
 }